refactor(iped-bfac): Extract nested classes from BfacApiClient to separate files
- Move LoginResult, ValidationResult, SendHashResult, SubmissionResult, and FileUploadStatus from BfacApiClient inner classes to standalone classes in iped.bfac.api package
- Update imports in BfacDialog and SubmissionWorker to reference new class locations
- Improve code organization and maintainability by separating API result types
